+++ lib/jsapi/meta/request_body/model.rb
@@ -9,6 +9,11 @@ module Jsapi
         delegate_missing_to :schema
 
         ##
+        # :attr: content_type
+        # The content type. <code>"application/json"</code> by default.
+        attribute :content_type, String, default: 'application/json'
+
+        ##
         # :attr: description
         # The optional description of the request body.
         attribute :description, String
@@ -25,7 +30,7 @@ module Jsapi
 
         def initialize(keywords = {})
           keywords = keywords.dup
-          super(keywords.extract!(:description, :examples))
+          super(keywords.extract!(:content_type, :description, :examples))
 
           add_example(value: keywords.delete(:example)) if keywords.key?(:example)
           keywords[:ref] = keywords.delete(:schema) if keywords.key?(:schema)
@@ -54,7 +59,7 @@ module Jsapi
           with_openapi_extensions(
             description: description,
             content: {
-              'application/json' => {
+              content_type => {
                 schema: schema.to_openapi(version),
                 examples: examples&.transform_values(&:to_openapi)
               }.compact